requirements.txt
dj-database-url==0.4.0
Django==1.9.5
django-cors-headers==1.1.0
django-extensions==1.6.1
django-oauth-toolkit==0.10.0
django-registration==2.0.4
django-rest-framework-social-oauth2==1.0.4
djangorestframework==3.3.2
djangorestframework-oauth==1.1.0
gunicorn==19.4.5
httpie==0.9.3
python-social-auth==0.2.14
rest-social-auth==0.4.3
psycopg2==2.6.1
whitenoise==3.0